MINIHACK-SEMANTIC-KERNEL/requirements.txt
semantic-kernel
python-dotenv
azure-search-documents
azure-identity
PyPDF2